% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_visualization.R
\name{bulk_group_input}
\alias{bulk_group_input}
\title{Format bulk groups input}
\usage{
bulk_group_input(bulkgroups, bulkidcol = "bulk_id", groupcol = "group")
}
\arguments{
\item{bulkgroups}{A vector or dataframe holding the bulk-group association.}

\item{bulkidcol}{If bulkgroups is a tibble, the name of the column that holds
the bulk ids.}

\item{groupcol}{If bulkgroups is a tibble, the name of the column that holds
the group ids.}
}
\value{
A named vector of group ids with corresponding bulk ids as names.
}
\description{
An input wrapper for \code{density_distribution()} to accept both named vectors as
well as dataframes containing bulk groups.
}
\examples{
# Case 1
bulkgroups <- c("b1" = "group_1", "b2" = "group_2", "b3" = "group_3")

bulk_group_input_1 <- tissueResolver:::bulk_group_input(
   bulkgroups,
   bulkidcol = "bulk_id",
   groupcol = "group"
)

# Case 2
library(tibble)
bulkgroups <- tibble(
  "bulk_id" = c("b1", "b2", "b3"),
  "group" = c("group_1", "group_2", "group_3")
)

bulk_group_input_2 <- tissueResolver:::bulk_group_input(
   bulkgroups,
   bulkidcol = "bulk_id",
   groupcol = "group"
)

}
