% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_proportions.R
\name{cell_proportions}
\alias{cell_proportions}
\title{Compute cell type proportions from tissuemodel}
\usage{
cell_proportions(
  tissuemodel,
  mapping = NULL,
  by = "celltype",
  weight = "weight",
  bulk_id = "bulk_id",
  cell_id = "cell_id"
)
}
\arguments{
\item{tissuemodel}{A tissuemodel as returned by \code{fit_tissue()}, or a data
frame that contains cell ids, weights and a categorical variable, e.g.,
celltype.}

\item{mapping}{A dataframe that maps each cell (column name given by
\code{cell_id}) to a category (column name given by \code{by}).}

\item{by}{The name of the categorical variable defining groups of cells.
Defaults to \code{"celltype"}.}

\item{weight}{The name of the column that holds the fitted cell weights.
Defaults to \code{"weight"}.}

\item{bulk_id}{The name of the column that identifies the bulks.
Defaults to \code{"bulk_id"}.}

\item{cell_id}{The name of the column that identifies the cells.
Defaults to \code{"cell_id"}.}
}
\value{
A data frame holding the cumulated weights for each cell population
within each bulk for each bootstrap run as a list in the \code{weight_boot}
column, if the tissuemodel included bootstrapping. The \code{weight} column
holds the mean over the bootstrap runs across each bulk-celltype group
and the last column holds the percentage of each cell type within a bulk.
}
\description{
Returns the sum over weights of certain cell populations (grouped by
mapping) for each bulk for a \code{tissuemodel} retrieved by \code{fit_tissue()}.
}
\examples{
library(tibble)
set.seed(42)
ngenes <- 4
nbulks <- 3
ncells <- 5

bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))

rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

fitted_tissue_boot <- fit_tissue(
  bulks,
  sc,
  bootstrap = TRUE,
  bootstrap_nruns = 5,
  bootstrap_pctcells = 50
)

mapping <- tibble(
  cell_id = cellnames,
  celltype = c("CTA", "CTB", "CTA", "CTB", "CTA")
)

cell_proportions_boot <- cell_proportions(
  tissuemodel = fitted_tissue_boot,
  mapping = mapping,
  by = "celltype",
  weight = "weight",
  bulk_id = "bulk_id",
  cell_id = "cell_id"
)

}
