% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_input_dimensions}
\alias{check_input_dimensions}
\title{Checks if the number
of genes of bulk data and the single cell library match.}
\usage{
check_input_dimensions(y, x)
}
\arguments{
\item{y}{matrix representing the single cell library}

\item{x}{matrix representing the collection of bulk samples}
}
\description{
Checks if the number
of genes of bulk data and the single cell library match.
}
\examples{
set.seed(1)
ngenes <- 4
nbulks <- 3
ncells <- 5
bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif((ngenes + 1) * ncells), nrow = ngenes + 1, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes + 1), seq.int(1, ngenes + 1))
rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

common <- tissueResolver:::common_genes(bulks, sc)
bulks_common <- common$A
sc_common <- common$B
tissueResolver:::check_input_dimensions(bulks_common, sc_common)


}
