% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_input_mapping}
\alias{check_input_mapping}
\title{checks if the mapping (mapping cell_ids to celltypes) is valid}
\usage{
check_input_mapping(tissuemodel, mapping, by, cell_id)
}
\arguments{
\item{tissuemodel}{tissuemodel as returned by fit_tissue}

\item{mapping}{dataframe that maps cell_ids to celltypes}

\item{by}{the name of the column that categorizes the mapping dataframe}

\item{cell_id}{name of the column in the mapping dataframe that names the single cells}
}
\value{
nothing
set.seed(1)
ngenes <- 4
nbulks <- 3
ncells <- 5
bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

fitted_tissue <- fit_tissue_noboot(bulks, sc)

mapping <- tibble(
cell_id = cellnames,
celltype = c("CTA", "CTB", "CTA", "CTB", "CTA")
)

check_input_mapping(fitted_tissue, mapping, by = "cell_id", cell_id = "cell_id")
}
\description{
checks if the mapping dataframe is a valid mapping for the cells in tissuemodel
}
