% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_input_se}
\alias{check_input_se}
\title{checks if the input for fast_specific_expression_regulation is valid}
\usage{
check_input_se(celldf, sclibrary, by, weight, bulk_id, cell_id)
}
\arguments{
\item{celldf}{dataframe composed of tissuemodel returned by fit_tissue and mapping prescribing the mapping of cell_ids to cell types}

\item{sclibrary}{single cell data matrix, columns are individual cells and must be indexable by the cell_ids contained in the celldf}

\item{by}{name of the categorical variable defining groups of cells}

\item{weight}{name of the column that holds the fitted cell weights}

\item{bulk_id}{name of the column that identifies the bulks}

\item{cell_id}{name of the column that contains the ids for the cells that match the column names of sclibrary}
}
\value{
nothing
}
\description{
checks if the input for computing specific expression is valid (accessed by fast_specific_expression_regulation)
}
\examples{
set.seed(1)
ngenes <- 4
nbulks <- 3
ncells <- 5

bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))

rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

fitted_tissue <- fit_tissue(bulks,sc)

mapping <- tibble(
  cell_id = cellnames,
  celltype = c("CTA", "CTB", "CTA", "CTB", "CTA")
)

celldf <- fitted_tissue$tissuemodel \%>\% inner_join(mapping, by = "cell_id")

tissueResolver:::check_input_se(
  celldf = celldf,
  sclibrary = sc,
  by = "celltype",
  weight = "weight",
  bulk_id = "bulk_id",
  cell_id = "cell_id"
)
}
