% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_visualization.R
\name{density_distribution}
\alias{density_distribution}
\title{Get density distributions of tissuemodels}
\usage{
density_distribution(
  scembedding,
  x = "UMAP1",
  y = "UMAP2",
  scids = "cell_id",
  tissuemodel,
  model_cellids = "cell_id",
  model_bulkids = "bulk_id",
  bulkgroups = NULL,
  bulk_groupids = "group",
  verbose = TRUE
)
}
\arguments{
\item{scembedding}{A dataframe that holds the 2D embeddings of the single
cell data used to compute the tissue model.}

\item{x}{The name of the x coordinate of the 2D embedding. Defaults to
\code{"UMAP1"}.}

\item{y}{The name of the y coordinate of the 2D embedding. Defaults to
\code{"UMAP2"}.}

\item{scids}{The name of the column holding the cell identifiers in the
embedding data frame. Defaults to \code{"cell_id"}.}

\item{tissuemodel}{A tissuemodel derived from \code{fit_tissue()}.}

\item{model_cellids}{The name of the cell_id column in the model data frame.
Defaults to \code{"cell_id"}.}

\item{model_bulkids}{The name of the bulk id column in the model data frame.
Defaults to \code{"bulk_id"}.}

\item{bulkgroups}{An optional mapping for the bulks to additional groups
(e.g. treated/untreated, etc).}

\item{bulk_groupids}{The names of the column that holds the group id within
bulkgroups if that is a dataframe. Defaults to \code{"group"}.}

\item{verbose}{Prints more output if \code{TRUE} (default).}
}
\value{
A dataframe with a kde density z-value for every x and y coordinate
and an additional group column.
}
\description{
Computes the weighted 2D densities for cell populations in a tissuemodel
derived from \code{fit_tissue()}. If groups are given, the group-specific
densities are computed.
}
\examples{
library(tibble)
set.seed(42)
scembedding <- data.frame(
  "colnames" = paste0("cell_", 1:9),
  "UMAP1" = c(rep(x = 1, n = 3), rep(x = 2, n = 3), rep(x = 3, n = 3)),
  "UMAP2" = c(rep(x = 1:3, each = 3)),
  "c.type" = sample(c("CTA", "CTB", "CTC"), 9, replace = TRUE)
)

fitted_tissue <- tibble(
  "fitted_genes" = NA,
  "tissuemodel" = tibble(
    "bulk_id" = c("b1", "b1", "b1", "b2", "b2", "b2", "b3", "b3", "b3"),
    "cell_id" = c(
      paste0("cell_", 1:3), paste0("cell_", 4:6), paste0("cell_", 1:3)
    ),
    "weight" = c(1, 0.5, 0.25, 0.25, 0.5, 1, 0.3, 0.3, 0.3)
  )
)

bulkgroups <- c("b1" = "bgroup_1", "b2" = "bgroup_2", "b3" = "bgroup_1")

density <- density_distribution(
  scembedding = scembedding,
  x = "UMAP1", y = "UMAP2", scids = "colnames",
  fitted_tissue$tissuemodel,
  bulkgroups = bulkgroups
)

}
