% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{df_to_matrix}
\alias{df_to_matrix}
\title{Convert a tibble or dataframe to a matrix}
\usage{
df_to_matrix(df, rows = "gene", cols = "cell_id", values = "expression")
}
\arguments{
\item{df}{A dataframe to be converted into matrix.}

\item{rows}{The name in the input dataframe that corresponds to the rows of
the resulting matrix, defaults to \code{"gene"}.}

\item{cols}{The name in the input dataframe that corresponds to the columns
of the resulting matrix, defaults to \code{"cell_id"}.}

\item{values}{The name in the input dataframe that will hold the entries of
the resulting matrix, defaults to \code{"expression"}.}
}
\value{
A matrix storing the values of \code{df}.
}
\description{
Convert a tibble or dataframe to a matrix to allows to call functions
expecting matrices to be called with tibbles.
}
\examples{
set.seed(42)
ngenes <- 4
nbulks <- 3
ncells <- 5
bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

fitted_tissue <- fit_tissue(bulks, sc)

sc_df <- tissueResolver:::matrix_to_df(
  sc,
  rows = "gene",
  cols = "cell_id",
  values = "expression"
)

sc_mat <- tissueResolver:::df_to_matrix(
  sc_df,
  rows = "gene",
  cols = "cell_id",
  values = "expression"
)

}
