% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_expression.R
\name{explained_expression}
\alias{explained_expression}
\title{Compute TOTAL explained expression for quality control}
\usage{
explained_expression(
  tissuemodel,
  sclibrary,
  by = "celltype",
  weight = "weight",
  bulk_id = "bulk_id",
  cell_id = "cell_id"
)
}
\arguments{
\item{tissuemodel}{A tissuemodel (tibble) as returned by \code{fit_tissue()}}

\item{sclibrary}{A single cell data matrix, columns are individual cells and
must be indexable by \code{cell_id} contained in the \code{tissuemodel}.}

\item{by}{The name of the categorical variable defining the groups of cells.
Defaults to \code{"celltype"}.}

\item{weight}{The name of the column that holds the fitted cell weights.
Defaults to \code{"weight"}.}

\item{bulk_id}{The name of the column that identifies the bulks. Defaults to
\code{"bulk_id"}.}

\item{cell_id}{The name of the column that contains the IDs for the cells
matching the column names of \code{sclibrary}. Defaults to \code{"cell_id"}.}
}
\value{
A tibble holding the explained expression for each gene in each bulk.
}
\description{
Computes the TOTAL explained expression, i.e., the specific gene expression
within each bulk. I.e., it uses the weights computed in fit tissue to
determine the explained specific expression of each gene in each bulk sample
without grouping cells.
}
\examples{
library(tibble)
set.seed(42)
ngenes <- 4
nbulks <- 3
ncells <- 5

bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))

rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

fitted_tissue <- fit_tissue(bulks, sc)

mapping <- tibble(
  cell_id = cellnames,
  celltype = c("CTA", "CTB", "CTA", "CTB", "CTA")
)

csre_explained <- explained_expression(
            tissuemodel=fitted_tissue,
            sclibrary=sc,
            by="celltype",
            weight="weight",
            bulk_id = "bulk_id",
            cell_id="cell_id")

}
