% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_expression.R
\name{fast_specific_expression_regulation}
\alias{fast_specific_expression_regulation}
\title{Compute the specific expression and regulation from a tissuemodel}
\usage{
fast_specific_expression_regulation(
  tissuemodel,
  sclibrary,
  mapping,
  by = "celltype",
  weight = "weight",
  bulk_id = "bulk_id",
  cell_id = "cell_id",
  compute_total = TRUE,
  show_pb = TRUE
)
}
\arguments{
\item{tissuemodel}{A tissuemodel (tibble) as returned by ´fit_tissue()´}

\item{sclibrary}{A single cell data matrix, columns are individual cells and
must be indexable by ´cell_id´ contained in the ´tissuemodel´.}

\item{mapping}{A dataframe that maps each cell (column name given by
´cell_id´) to a category (column name given by ´by´).}

\item{by}{The name of the categorical variable defining the groups of cells.
Defaults to ´"celltype"´.}

\item{weight}{The name of the column that holds the fitted cell weights.
Defaults to ´"weight"´.}

\item{bulk_id}{The name of the column that identifies the bulks. Defaults to
´"bulk_id"´.}

\item{cell_id}{The name of the column that contains the IDs for the cells
matching the column names of ´sclibrary´. Defaults to ´"cell_id"´.}

\item{compute_total}{If set to ´TRUE´, also compute the total expression
explained by ALL cell types. Can be used to compare resulting virtual
tissue to the bulk.}

\item{show_pb}{If \code{FALSE}, the function will not display a progress bar.
Defaults to \code{TRUE}.}
}
\value{
A tibble holding the expression and regulation of each gene within
each bulk and celltype.
}
\description{
Computes the specific expression and regulation for a certain group of
cells, weighting the individual cells accordingly. Normalization of the
expression column is such that the total expression of a gene is
approximately that of the bulk, while the regulation column corresponds to
the average expression of a single cell of that type. This function provides
the main computations accessed by ´specific_expression_regulation()´
}
\examples{
library(dplyr)
library(tibble)
set.seed(42)
ngenes <- 4
nbulks <- 3
ncells <- 5

bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))

rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

fitted_tissue <- fit_tissue(bulks, sc)

mapping <- tibble(
  cell_id = cellnames,
  celltype = c("CTA", "CTB", "CTA", "CTB", "CTA")
)

celldf <- fitted_tissue$tissuemodel \%>\% inner_join(mapping, by = "cell_id")

csre <- tissueResolver:::fast_specific_expression_regulation(
  tissuemodel = fitted_tissue,
  sclibrary = sc,
  mapping = mapping,
  by = "celltype",
  weight = "weight",
  bulk_id = "bulk_id",
  cell_id = "cell_id",
  compute_total = TRUE
)

}
