% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tissue.R
\name{fit_tissue}
\alias{fit_tissue}
\title{Fits a given single cell library to a collection of bulk samples by
weighting the single cell gene expressions
(the columns of the library matrix) in order to
optimally match the bulk gene expressions.
Allows for bootstrapping specified percentage of cells.}
\usage{
fit_tissue(
  bulkdata,
  sclibrary,
  bootstrap = FALSE,
  bootstrap_nruns = 50,
  bootstrap_pctcells = 10
)
}
\arguments{
\item{bulkdata}{dataframe containing the gene expression of
a bulk sample in each column. Rownames are expected to be gene identifieres}

\item{sclibrary}{dataframe containing a single cell profile in
each column. Rownames are expected to be gene identifieres}

\item{bootstrap}{if set to FALSE fit tissue without bootstrapping.
If set to true select bootstrap_pctcells perecent of cells
and perform bootstrap_nruns of sampling cells and fitting these to bulks}

\item{bootstrap_nruns}{number of bootstrap runs}

\item{bootstrap_pctcells}{integer in \link{0,100} which prescribes the percentage of cells
sampled from the single cell library in each bootstrap run}
}
\value{
if bootstrap is FALSE, returns a list with entries
\itemize{
\item tissuemodel: a data frame which holds all weights for each ' selected cell (id) and the bulk id that it was fitted to.
\item fitted_genes: a list of fitted gene names
\item bootstrap: FALSE
If bootstrap is TRUE it returns a list containing
\item tissuemodels: a list of tissuemodels of every bootstrap run
\item fitted_genes: a list of fitted gene names
\item result$tissuemodels[\link{irun}] <- this_model$tissuemodel
\item bootstrap: TRUE
\item bootstrap_nruns: number of bootstrap samples
\item bootstrap_pctcells: percentage of all single cells that were used in every bootstrap run
}
}
\description{
Fits a given single cell library to a collection of bulk samples by
weighting the single cell gene expressions
(the columns of the library matrix) in order to
optimally match the bulk gene expressions.
Allows for bootstrapping specified percentage of cells.
}
\examples{
set.seed(1)
ngenes <- 4
nbulks <- 3
ncells <- 5

bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))

rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

fitted_tissue <- fit_tissue(bulks,sc, bootstrap = TRUE, bootstrap_nruns = 5, bootstrap_pctcells = 50)
}
