% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tissue.R
\name{fit_tissue_noboot}
\alias{fit_tissue_noboot}
\title{fits a given single cell library to a collection of bulk samples by
weighting the single cell gene expressions
(the columns of the library matrix) in order to
optimally match the bulk gene expressions
The bulks are fit in parallel on multiple cores}
\usage{
fit_tissue_noboot(bulkdata, sclibrary, maxit = 2000, ncores = 1)
}
\arguments{
\item{bulkdata}{dataframe containing the gene expression of
a bulk sample in each column. Rownames are expected to be gene identifieres.}

\item{sclibrary}{dataframe containing a single cell profile in
each column. Rownames are expected to be gene identifieres}

\item{maxit}{the iteration limit used in the optimization algorithm, defaults to 2e3}

\item{ncores}{the number of cores for parallel computation}
}
\value{
a list containing
-tissuemodel: contains the computed cell weights for each bulk sample
-fitted_genes: the gene names of fitted genes
-log_store: logging info for residual across iterations
}
\description{
fits a given single cell library to a collection of bulk samples by
weighting the single cell gene expressions
(the columns of the library matrix) in order to
optimally match the bulk gene expressions
The bulks are fit in parallel on multiple cores
}
\examples{
set.seed(1)
ngenes <- 4
nbulks <- 3
ncells <- 5

bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))

rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames
fitted_tissue <- tissueResolver:::fit_tissue_noboot(bulks, sc, 2e3, 2)

}
