% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_tissuemodel}
\alias{is_tissuemodel}
\title{Check if an object is a tissuemodel}
\usage{
is_tissuemodel(tissuemodel)
}
\arguments{
\item{tissuemodel}{object returned by \code{fit_tissue_noboot()} to be checked}
}
\value{
\code{TRUE} if it is a tissuemodel and \code{FALSE} if not.
}
\description{
Checks if the object is a tissuemodel as returned by \code{fit_tissue()}.
}
\examples{
set.seed(42)
ngenes <- 4
nbulks <- 3
ncells <- 5
bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

fitted_tissue <- fit_tissue(bulks, sc)

tissueResolver:::is_tissuemodel(fitted_tissue)

}
