% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{Convert a matrix to a tibble}
\usage{
matrix_to_df(
  genematrix,
  rows = "gene",
  cols = "cell_id",
  values = "expression"
)
}
\arguments{
\item{genematrix}{A matrix to convert.}

\item{rows}{The dedicated name in the resulting dataframe storing the row
names of the input matrix, defaults to \code{"gene"}.}

\item{cols}{The dedicated name in the resulting dataframe storing the column
names of the input matrix, defaults to \code{"cell_id"}.}

\item{values}{The dedicated name in the resulting dataframe that holds the
entries of the input matrix, defaults to \code{"expression"}.}
}
\value{
A tibble storing the matrix values and names.
}
\description{
Convert a matrix to a tibble to allows to call functions expecting tibbles
to be called with dataframes.
}
\examples{
set.seed(42)
ngenes <- 4
nbulks <- 3
ncells <- 5
bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

fitted_tissue <- fit_tissue(bulks, sc)

sc_df <- tissueResolver:::matrix_to_df(
   sc,
   rows = "gene",
   cols = "cell_id",
   values = "expression"
  )

sc_mat <- tissueResolver:::df_to_matrix(
   sc_df,
   rows = "gene",
   cols = "cell_id",
   values = "expression"
)

}
