% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cse_visualization.R
\name{plot_csre}
\alias{plot_csre}
\title{Plot the cell type specific expression between two bulk groups}
\usage{
plot_csre(
  csre,
  heatmapviz = "relchange",
  barplotviz = "linear",
  groupA = NA,
  groupB = NA,
  ctypes = NULL,
  addexpressionlevel = FALSE
)
}
\arguments{
\item{csre}{An object returned by \code{specific_expression_regulation()} with
an added \code{group} column that indicates the bulk group of each bulk.}

\item{heatmapviz}{The quantity to show in the heat map. Can be set to
\code{ "logdiffexpr"} (logarithmic differential expression), \code{"logdiffreg"}
(log differential regulation per cell), \code{"relchange"} (relative change
in expression compared to the average of the two groups), or
\code{"relregchange"} (relative change in regulation per cell compared to
the average of the two groups).}

\item{barplotviz}{The transformation to show in the bar plot for each gene
Can be set to:
\itemize{
\item \code{"linear"}: Show the reconstructed gene expression of the two bulk
groups as a stacked colored bar plot.
\item \code{"log"}: Show the total log-expression in both groups.
\item \code{"logdiff"}: Show the differential log-expression between the two
groups in a single bar plot.
\item \code{"relative"}: Show the group-A expression as 100\% fractions and the
group-B relative to that.
\item \code{"relativeB"}: Similarly, but normalized to the group-B expression.
}}

\item{groupA}{The identifier of bulk group A (factor in the \code{group} column).}

\item{groupB}{The identifier of bulk group B (factor in the \code{group} column).}

\item{ctypes}{If not \code{NULL}, the order in which the cell types will be
displayed. All cell types present in the dataset but not in this array
will be summed as \code{"other"}.}

\item{addexpressionlevel}{If set to \code{TRUE} and \code{barplotviz = "relative"},
add an additional bar plot that indicates the expression level of the
gene present in bulk group A.}
}
\value{
A ggplot object.
}
\description{
This function generates several bar and heatmap plots for comparing the
cell type specific expression between two bulk groups.
}
\examples{
library(tibble)
library(dplyr)
library(tidyr)
set.seed(42)
ngenes <- 4
nbulks <- 3
ncells <- 5

bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))

rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

mapping <- tibble(
  cell_id = cellnames,
  celltype = c("CTA", "CTB", "CTA", "CTC", "CTA")
)

grouping <- tibble(
  "bulk_id" = c("bulk_1", "bulk_2", "bulk_3"),
  "group" = c("bgroup_1", "bgroup_2", "bgroup_1")
)

fitted_tissue <- fit_tissue(
  bulks,
  sc
)

csre <- specific_expression_regulation(
  tissuemodel = fitted_tissue,
  sclibrary = sc,
  mapping = mapping,
  by = "celltype",
  weight = "weight",
  bulk_id = "bulk_id",
  cell_id = "cell_id",
  compute_total = TRUE
)

csre <- csre \%>\%
  inner_join(grouping \%>\% dplyr::select(bulk_id, group)) \%>\%
  drop_na()

csre_plot <- plot_csre(
  csre = csre,
  heatmapviz = "relchange",
  barplotviz = "relative",
  groupA = "bgroup_1",
  groupB = "bgroup_2",
  ctypes = c("CTA", "CTC"),
  addexpressionlevel = TRUE
)

print(csre_plot)

}
