% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_visualization.R
\name{plot_differential_densities}
\alias{plot_differential_densities}
\title{Plot differential densities between two groups}
\usage{
plot_differential_densities(
  densitydata,
  scembedding = NULL,
  groupA = NA,
  groupB = NA,
  embedding.x = "UMAP1",
  embedding.y = "UMAP2",
  colourby = NULL
)
}
\arguments{
\item{densitydata}{A dataframe returned by \code{density_distribution()}.}

\item{scembedding}{A dataframe or matrix that holds 2D embeddings of the
single cell data used to compute the tissuemodel. Defaults to \code{NULL}. If
\code{scembedding} is provided the single cell embedding is plotted.}

\item{groupA}{The name of the group with positive weights. These are colored
red, if weights are larger than in \code{groupB}.}

\item{groupB}{The name of the group with negative weights. These are colored
blue, if weights are larger than in \code{groupA}.}

\item{embedding.x}{The column name of the x coordinate for the 2D embedding
in \code{scembedding}. Defaults to \code{"UMAP1"}.}

\item{embedding.y}{The column name of the y coordinate for the 2D embedding
in \code{scembedding}. Defaults to \code{"UMAP2"}.}

\item{colourby}{defaults to \code{NULL}. If column name is provided, colour the
single cells by this column of \code{scembedding} (e.g., celltype).}
}
\value{
A ggplot plot object that can be further customized, viewed or saved
in a file. For more details see \code{ggplot2}.
}
\description{
Visualizes groups for which a tissue was fitted and their differences. The
difference in density between groupA and groupB is computed and plotted. If
group labels are omitted, they are guessed (if only two groups exist).
In addition, the single cell embedding and a contour plot of the density for
each bulk group are plotted.
}
\examples{
library(tibble)
library(ggplot2)
set.seed(42)
scembedding <- data.frame(
  "colnames" = paste0("cell_", 1:9),
  "UMAP1" = c(rep(x = 1, n = 3), rep(x = 2, n = 3), rep(x = 3, n = 3)),
  "UMAP2" = c(rep(x = 1:3, each = 3)),
  "c.type" = sample(c("CTA", "CTB", "CTC"), 9, replace = TRUE)
)
fitted_tissue <- tibble(
  "fitted_genes" = NA,
  "tissuemodel" = tibble(
    "bulk_id" = c("b1", "b1", "b1", "b2", "b2", "b2", "b3", "b3", "b3"),
    "cell_id" = c(
      paste0("cell_", 1:3), paste0("cell_", 4:6), paste0("cell_", 1:3)
    ),
    "weight" = c(1, 0.5, 0.25, 0.25, 0.5, 1, 0.3, 0.3, 0.3)
  )
)
bulkgroups <- c("b1" = "bgroup_1", "b2" = "bgroup_2", "b3" = "bgroup_1")
density <- density_distribution(
  scembedding = scembedding,
  x = "UMAP1", y = "UMAP2", scids = "colnames",
  fitted_tissue$tissuemodel,
  bulkgroups = bulkgroups
)

differential_densities <- plot_differential_densities(
  density,
  scembedding,
  groupA = "bgroup_1",
  groupB = "bgroup_2",
  embedding.x = "UMAP1",
  embedding.y = "UMAP2",
  colourby = "c.type"
)

print(differential_densities)

}
