% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_visualization.R
\name{plot_element_differential_density}
\alias{plot_element_differential_density}
\title{Plot the differential density between two groups as a tile plot}
\usage{
plot_element_differential_density(
  densities,
  groupA = NA,
  groupB = NA,
  x = "x",
  y = "y",
  z = "z",
  group = "group"
)
}
\arguments{
\item{densities}{A dataframe returned by \code{density_distribution()}.}

\item{groupA}{The name of the group with positive weights. These are colored
red, if weights are larger than in \code{groupB}.}

\item{groupB}{The name of the group with negative weights. These are colored
blue, if weights are larger than in \code{groupA}.}

\item{x}{The column name of the x coordinate of the kde grid.}

\item{y}{The column name of the y coordinate of the kde grid.}

\item{z}{The column name of the z coordinate for the kde grid.}

\item{group}{The column name of the densities dataframe storing the
bulkgroups.}
}
\value{
A ggplot plot object that can be further customized, viewed or saved
in a file. For more details see \code{ggplot2}.
}
\description{
The kde density at each grid point of the single cell embedding for each
bulkgroup provided by \code{density_distribution()} is used to assess the
difference in densities between two groups. This differential density is
then visualized in the kde grid space as tile plot. If group labels are
omitted, they are guessed (if only two groups exist).
}
\examples{
library(tibble)
library(ggplot2)
set.seed(42)
scembedding <- data.frame(
  "colnames" = paste0("cell_", 1:9),
  "UMAP1" = c(rep(x = 1, n = 3), rep(x = 2, n = 3), rep(x = 3, n = 3)),
  "UMAP2" = c(rep(x = 1:3, each = 3)),
  "c.type" = sample(c("CTA", "CTB", "CTC"), 9, replace = TRUE)
)
fitted_tissue <- tibble(
  "fitted_genes" = NA,
  "tissuemodel" = tibble(
    "bulk_id" = c("b1", "b1", "b1", "b2", "b2", "b2", "b3", "b3", "b3"),
    "cell_id" = c(
      paste0("cell_", 1:3), paste0("cell_", 4:6), paste0("cell_", 1:3)
    ),
    "weight" = c(1, 0.5, 0.25, 0.25, 0.5, 1, 0.3, 0.3, 0.3)
  )
)
bulkgroups <- c("b1" = "bgroup_1", "b2" = "bgroup_2", "b3" = "bgroup_1")
density <- density_distribution(
  scembedding = scembedding,
  x = "UMAP1", y = "UMAP2", scids = "colnames",
  fitted_tissue$tissuemodel,
  bulkgroups = bulkgroups
)

element_differential_density <-
  tissueResolver:::plot_element_differential_density(
    density,
    groupA = "bgroup_1",
    groupB = "bgroup_2",
    x = "x",
    y = "y",
    z = "z",
    group = "group"
)

print(ggplot() +
  element_differential_density +
  coord_fixed())

}
