% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_visualization.R
\name{plot_element_embedding}
\alias{plot_element_embedding}
\title{Plots the single cell embedding}
\usage{
plot_element_embedding(scembedding, x = "UMAP1", y = "UMAP2", colourby = NULL)
}
\arguments{
\item{scembedding}{A dataframe that holds 2D embeddings of the single cell
data used to compute the tissuemodel.}

\item{x}{The column name of the x coordinate for the 2D embedding in
\code{scembedding}.}

\item{y}{The column name of the y coordinate for the 2D embedding in
\code{scembedding}.}

\item{colourby}{If not \code{NULL}, colour the single cells by this column of
\code{scembedding} (e.g. celltype).}
}
\value{
A ggplot plot object that can further be customized, viewed or saved
in a file. For more details see \code{ggplot2}.
}
\description{
Plots each UMAP point from the single cell embedding. These may be coloured
(e.g. by the associated cell type).
}
\examples{
library(ggplot2)
scembedding <- data.frame(
  "colnames" = paste0("cell_", 1:9),
  "UMAP1" = c(rep(x = 1, n = 3), rep(x = 2, n = 3), rep(x = 3, n = 3)),
  "UMAP2" = c(rep(x = 1:3, each = 3)),
  "c.type" = sample(c("CTA", "CTB", "CTC"), 9, replace = TRUE)
)

embedding_plot <- tissueResolver:::plot_element_embedding(
  scembedding,
  x = "UMAP1",
  y = "UMAP2",
  colourby = "c.type"
)

print(ggplot() +
  embedding_plot +
  coord_fixed())

}
