% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{progress_bar}
\alias{progress_bar}
\title{Set up a progress bar}
\usage{
progress_bar(total, min = 0, style = 3, ...)
}
\arguments{
\item{total}{The total number of steps for the progress bar.}

\item{min}{The starting value for the progress bar (defaults to \code{0}).}

\item{style}{The style of the progress bar (defaults to \code{3}).}

\item{...}{Additional arguments passed to \code{utils::txtProgressBar()}.}
}
\value{
A function to update the progress bar.
}
\description{
Initializes a progress bar and returns a function to update it.
}
