% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{promote_matrix}
\alias{promote_matrix}
\title{Promote a vector to a matrix}
\usage{
promote_matrix(vec)
}
\arguments{
\item{vec}{A matrix or vector.}
}
\value{
Promoted matrix.
}
\description{
Promotes a vector to a matrix. If the input is already a matrix, this matrix
is returned.
}
