% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_visualization.R
\name{prop_visualization}
\alias{prop_visualization}
\title{Plot the cell type proportions between two groups}
\usage{
prop_visualization(props, grouping, groupA, groupB)
}
\arguments{
\item{props}{A dataframe containing cell type proportions (output of
\code{cell_proportions()}).}

\item{grouping}{A dataframe containing \code{bulk_id} and \code{group} columns for
group association of samples.}

\item{groupA}{Group that is in the numerator of the ratio.}

\item{groupB}{Group that is in the denominator of the ratio.}
}
\value{
A ggplot object.
}
\description{
This function visualizes the ratio of cell type proportions between two
groups based on the output of \code{cell_proportions()} and bulk grouping data.
}
