% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_scores.R
\name{quality_scores}
\alias{quality_scores}
\title{Compute quality scores for genes and bulks}
\usage{
quality_scores(csre, bulks)
}
\arguments{
\item{csre}{The cell type specific expression (\code{csre}) obtained by
\code{specific_expression_regulation()}. Note that \code{csre} must include the
total explained gene expression to calculate the quality scores.}

\item{bulks}{A matrix or dataframe containing the gene expression of bulk
samples in each column and genes in each row. Row names must match the
gene names in \code{csre}.}
}
\value{
A list of dataframes:
\itemize{
\item \code{genes}: Contains quality scores by gene + mean gene-wise actual and
explained total expression
\item \code{bulks}: Contains quality scores to judge individual virtual tissues
}
}
\description{
This function computes gene-specific and sample-specific quality scores.
Specifically, it computes the mean relative residual and average bootstrap
variance for each gene and each bulk sample.
}
\examples{
library(tibble)
set.seed(42)
ngenes <- 4
nbulks <- 3
ncells <- 5
bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames
mapping <- tibble(
  cell_id = cellnames,
  celltype = c("CTA", "CTB", "CTA", "CTC", "CTA")
)
grouping <- tibble(
  "bulk_id" = c("bulk_1", "bulk_2", "bulk_3"),
  "group" = c("bgroup_1", "bgroup_2", "bgroup_1")
)
fitted_tissue <- fit_tissue(
  bulks,
  sc,
  bootstrap = TRUE,
  bootstrap_nruns = 5,
  bootstrap_pctcells = 50
)
csre <- specific_expression_regulation(
  tissuemodel = fitted_tissue,
  sclibrary = sc,
  mapping = mapping,
  by = "celltype",
  weight = "weight",
  bulk_id = "bulk_id",
  cell_id = "cell_id",
  compute_total = TRUE
)

qc <- quality_scores(csre, bulks)

}
