% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{common_genes}
\alias{common_genes}
\title{Restricts rows of two matrices to the set of common genes
and brings them in the same ordering.
Additionally it will report if less than 20\% of the genes are in common}
\usage{
common_genes(A, B)
}
\arguments{
\item{A}{matrix A}

\item{B}{matrix B}
}
\value{
A list containing the input A and B
but restricted to the set of common genes
}
\description{
Restricts rows of two matrices to the set of common genes
and brings them in the same ordering.
Additionally it will report if less than 20\% of the genes are in common
}
\examples{
set.seed(1)
ngenes <- 4
nbulks <- 3
ncells <- 5
bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif((ngenes + 1) * ncells), nrow = ngenes + 1, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes + 1), seq.int(1, ngenes + 1))
rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

common <- tissueResolver:::common_genes(bulks, sc)
bulks_common <- common$A
sc_common <- common$B


}
