% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{df_to_matrix}
\alias{df_to_matrix}
\title{converts a tibble / dataframe to a matrix}
\usage{
df_to_matrix(df, rows = "gene", cols = "cell_id", values = "expression")
}
\arguments{
\item{df}{dataframe to be converted into matrix}

\item{rows}{name in the input dataframe that corresponds to the rows of the resulting matrix,
defaults to gene}

\item{cols}{name in the input dataframe that corresponds to the columns of the resulting matrix,
defaults to cell_id}

\item{values}{name in the input dataframe that will hold the entries of the resulting matrix,
defaults to expression}
}
\value{
a matrix storing the dataframe's values
}
\description{
allows to call functions expecting matrices to be called with tibbles
}
\examples{

set.seed(1)
ngenes <- 4
nbulks <- 3
ncells <- 5
bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

fitted_tissue <- fit_tissue(bulks, sc)

sc_df <- tissueResolver:::matrix_to_df(sc, rows = "gene", cols = "cell_id", values = "expression")

sc_mat <- tissueResolver:::df_to_matrix(sc_df, rows = "gene", cols = "cell_id", values = "expression")

}
