% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_tissuemodel}
\alias{is_tissuemodel}
\title{checks tissuemodel to have the desired format}
\usage{
is_tissuemodel(tissuemodel)
}
\arguments{
\item{tissuemodel}{object returned by fit_tissue_noboot to be checked}
}
\value{
TRUE if it is a tissuemodel, FALSE otherwise
}
\description{
checks if the object is a tissuemodel as returned by fit_tissue_noboot
}
\examples{
set.seed(1)
ngenes <- 4
nbulks <- 3
ncells <- 5
bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

fitted_tissue <- fit_tissue(bulks, sc)

tissueResolver:::is_tissuemodel(fitted_tissue)


}
