% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_visualization.R
\name{kde2d_weighted}
\alias{kde2d_weighted}
\title{weighted 2D Kernel density estimation}
\usage{
kde2d_weighted(
  x,
  y,
  weight,
  bandwidth = NULL,
  n.grid = 25,
  lims = c(range(x), range(y))
)
}
\arguments{
\item{x}{numeric, x coordinate of the embedding}

\item{y}{numeric, y coordinate of the embedding}

\item{weight}{numeric, weight for each data point}

\item{bandwidth}{bandwidth for the KDE, inducing the smoothing for the density estimate.
Numeric with length of 2,
defaults to 'c(MASS::bandwidth.nrd(x), MASS::bandwidth.nrd(y))/4'}

\item{n.grid}{integer, number of grid points for plotting, defaults to 25}

\item{lims}{numeric with length of 4 c(lower(x),upper(x), lower(y), upper(y)), range of the plot, allowing for plotting only a subframe of the data.
Defaults to 'c(range(x), range(y))'}
}
\value{
data.frame with 3 entries: 'x' and 'y' grid coordinates
in long format.
'z' (= weighted density) in long format
}
\description{
The following function is an adaption from MASS::kde2d,
}
\examples{
rm(list = ls()); gc()
library(tissueResolver)
library(ggplot2)

set.seed(1)
n <- 10
tmp.frame <- data.frame(
  x = c(rep(x = 1, n = 3), rep(x = 2, n = 3), rep(x = 3, n = 3)), 
  y = c(rep(x = 1:3, each = 3)), 
  c.type = sample(c("A", "B", "C"), 9, replace = TRUE), 
  group_0 = c(1, 0.5, 0.25,  rep(0, 6)),
  group_1 = c(rep(0, 3), rep(1, 3), rep(0, 3)), 
  group_2 = rev(c(rep(1, 3), rep(0, 6)))
)

weighted.density <- tissueResolver:::kde2d_weighted(
  x = tmp.frame$x
  , y = tmp.frame$y
  , weight = tmp.frame$group_0
)
# Note that the corresponding weight matrix is given by replicating
# the t(weight) across all rows of the grid

p <- ggplot(
  data = tmp.frame
  , mapping = aes(
    x = x
    , y = y
    )
  ) + 
  geom_tile(
    data = weighted.density
    , mapping = aes(
      x = x
      , y = y
      , fill = z
    )
  ) + 
  geom_point()
print(p)

}
