% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{converts a matrix to a tibble}
\usage{
matrix_to_df(
  genematrix,
  rows = "gene",
  cols = "cell_id",
  values = "expression"
)
}
\arguments{
\item{genematrix}{matrix to convert}

\item{rows}{dedicated name in the resulting dataframe storing the row names of the input matrix,
defaults to gene}

\item{cols}{dedicated name in the resulting dataframe storing the coulmn names of the input matrix,
default to cel_id}

\item{values}{name in the resulting dataframe that holds the entries of the input matrix,
defaults to expression}
}
\value{
a tibble storing the matrix values and names
}
\description{
allows to call functions expecting tibbles to be called with dataframes
}
\examples{
set.seed(1)
ngenes <- 4
nbulks <- 3
ncells <- 5
bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

fitted_tissue <- fit_tissue(bulks, sc)

sc_df <- tissueResolver:::matrix_to_df(sc, rows = "gene", cols = "cell_id", values = "expression")

sc_mat <- tissueResolver:::df_to_matrix(sc_df, rows = "gene", cols = "cell_id", values = "expression")

}
