% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cse_visualization.R
\name{plot_csre}
\alias{plot_csre}
\title{plots cell type specific expression comparing two bulk groups}
\usage{
plot_csre(
  csre,
  heatmapviz = "relchange",
  barplotviz = "linear",
  groupA = NA,
  groupB = NA,
  ctypes = NULL,
  addexpressionlevel = FALSE
)
}
\arguments{
\item{csre}{object returned by specific_expression_regulation, with an added bulk group column}

\item{heatmapviz}{Quantity to show in the heat map.
May be either "logdiffexpr" (for the logarithmic differential expression),
"logdiffreg" (log differential regulation, per cell), or
"relchange" (relative change in expression compared to the average of the two groups), or
"relregchange" (relative change in regulation -- per cell -- compared to the average of the two groups)}

\item{barplotviz}{transformation to show in the bar plot for each gene
May be either
"linear", which will show the reconstructed gene expression of the two bulk groups as a stacked colored bar plot, or
"log", showing the total log-expression in both groups, or
"logdiff", showing the differential log-expression between the two groups in a single bar plot,
"relative", or "relativeA" showing the group-A expression as 100\% fractions and the group-B relative to that.
"relativeB", similarly, but normalized to the group-B expression}

\item{groupA}{identifier of bulk group A (factor in the group column)}

\item{groupB}{identifier of bulk group B (factor in the group column)}

\item{ctypes}{if non-NULL, the order in which the cell types will be displayed.
All celltypes present in the dataset but not in this array will be summed as "other".}

\item{addexpressionlevel}{if set to TRUE and barplotviz == relative, add an additional bar plot that indicates the expression level of the gene present in bulk group A}
}
\value{
ggplot object
}
\description{
This function generates several bar and heatmapplots for comparing cell type specific expression comparing two bulk groups.
}
\examples{
library(gridExtra)
set.seed(1)
ngenes <- 4
nbulks <- 3
ncells <- 5

bulks <- matrix(runif(ngenes * nbulks), nrow = ngenes, ncol = nbulks)
sc <- matrix(runif(ngenes * ncells), nrow = ngenes, ncol = ncells)
bulknames <- paste0(rep("bulk_", nbulks), seq.int(1, nbulks))
cellnames <- paste0(rep("cell_", ncells), seq.int(1, ncells))
genenames_bulks <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))
genenames_sc <- paste0(rep("gene_", ngenes), seq.int(1, ngenes))

rownames(bulks) <- genenames_bulks
rownames(sc) <- genenames_sc
colnames(bulks) <- bulknames
colnames(sc) <- cellnames

mapping <- tibble(
  cell_id = cellnames,
  celltype = c("CTA", "CTB", "CTA", "CTC", "CTA")
)

grouping <- tibble(
  "bulk_id" = c("bulk_1", "bulk_2", "bulk_3"),
  "group" = c("bgroup_1", "bgroup_2", "bgroup_1")
)

fitted_tissue <- fit_tissue(
  bulks,
  sc
)

csre <- specific_expression_regulation(
  tissuemodel = fitted_tissue,
  sclibrary = sc,
  mapping = mapping,
  by = "celltype",
  weight = "weight",
  bulk_id = "bulk_id",
  cell_id = "cell_id",
  compute_total = TRUE
)

csre <- csre \%>\%
  inner_join(grouping \%>\% dplyr::select(bulk_id, group)) \%>\%
  drop_na()


csre_plot <- plot_csre(
  csre = csre,
  heatmapviz = "relchange",
  barplotviz = "relative",
  groupA = "bgroup_1",
  groupB = "bgroup_2",
  ctypes = c("CTA", "CTC"),
  addexpressionlevel = TRUE
)

print(csre_plot)
}
