% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_visualization.R
\name{plot_differential_densities}
\alias{plot_differential_densities}
\title{plots differential group densities between two groups A and B.}
\usage{
plot_differential_densities(
  densitydata,
  scembedding = NULL,
  groupA = NA,
  groupB = NA,
  embedding.x = "UMAP1",
  embedding.y = "UMAP2",
  colourby = NULL
)
}
\arguments{
\item{densitydata}{dataframe that is returned by density_distribution (see the help of that function)}

\item{scembedding}{embedding of the cells that the tissue model was computed on (dataframe or matrix-like). Optional. If NULL, no embedding will be plotted but just the densities.}

\item{groupA}{name of the group with the positive weights (appearing red in the plot where this group has a larger weight than groupB).}

\item{groupB}{name of the group with the negative relative weights (appearing blue in the plot where this group has a larger weight than groupA).}

\item{embedding.x}{column name of the x coordinate for the 2D embedding in scembedding}

\item{embedding.y}{column name of the y coordinate for the 2D embedding in scembedding}

\item{colourby}{if non-null, colour the single cells by this column of scembedding (e.g., celltype)}
}
\value{
a ggplot plot object that can further be customized, viewed or saved in a file. see ggplot2.
}
\description{
visualizes groups for which a tissue was fitted and their differences.
The difference in density between groupA and groupB is computed and plotted.
If group labels are omitted, they are tried to be guessed (if only two groups exist).
Besides the differential densities the plot contains the visualization of the single cell embedding
and a contour plot illustrating the density of each bulk group
}
\examples{
set.seed(1)
scembedding <- data.frame(
  "colnames" = paste0("cell_", 1:9),
  # "colnames" = c(paste0("cell_", 1:3), paste0("cell_", 1:3), paste0("cell_", 1:3)),
  "UMAP1" = c(rep(x = 1, n = 3), rep(x = 2, n = 3), rep(x = 3, n = 3)),
  "UMAP2" = c(rep(x = 1:3, each = 3)),
  "c.type" = sample(c("CTA", "CTB", "CTC"), 9, replace = TRUE)
)
fitted_tissue <- tibble(
  "fitted_genes" = NA,
  "tissuemodel" = tibble(
    "bulk_id" = c("b1", "b1", "b1", "b2", "b2", "b2", "b3", "b3", "b3"),
    # "cell_id" = paste0("cell_", 1:9),
    "cell_id" = c(paste0("cell_", 1:3), paste0("cell_", 4:6), paste0("cell_", 1:3)),
    "weight" = c(1, 0.5, 0.25, 0.25, 0.5, 1, 0.3, 0.3, 0.3)
  )
)
bulkgroups <- c("b1" = "bgroup_1", "b2" = "bgroup_2", "b3" = "bgroup_1")
density <- density_distribution(
  scembedding = scembedding,
  x = "UMAP1", y = "UMAP2", scids = "colnames",
  fitted_tissue$tissuemodel,
  bulkgroups = bulkgroups
)

differential_densities <- plot_differential_densities(
  density,
  scembedding,
  groupA = "bgroup_1",
  groupB = "bgroup_2",
  embedding.x = "UMAP1",
  embedding.y = "UMAP2",
  colourby = "c.type"
)

print(differential_densities)
}
