% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_visualization.R
\name{plot_element_differential_density}
\alias{plot_element_differential_density}
\title{plots the differential density between two groups as a tile plot}
\usage{
plot_element_differential_density(
  densities,
  groupA = NA,
  groupB = NA,
  x = "x",
  y = "y",
  z = "z",
  group = "group"
)
}
\arguments{
\item{densities}{dataframe that is returned by density_distribution (see the help of that function)}

\item{groupA}{name of the group with the positive weights (appearing red in the plot where this group has a larger weight than groupB).}

\item{groupB}{name of the group with the negative relative weights (appearing blue in the plot where this group has a larger weight than groupA).}

\item{x}{column name of the x coordinate of the kde grid}

\item{y}{column name of the y coordinate of the kde grid}

\item{z}{column name of the z coordinate for the kde grid}

\item{group}{column name of the densities dataframe storing the bulkgroups}
}
\value{
a ggplot plot object that can further be customized, viewed or saved in a file. see ggplot2.
}
\description{
having the kde density at each grid point of the single cell embedding for each bulkgroup provided by density_distribution, we
can assess the difference in densities between two of these groups. This differential densitiy is then visualized in the kde grid space as tile plot.
If group labels are omitted, they are tried to be guessed (if only two groups exist).
}
\examples{
set.seed(1)
scembedding <- data.frame(
  "colnames" = paste0("cell_", 1:9),
  # "colnames" = c(paste0("cell_", 1:3), paste0("cell_", 1:3), paste0("cell_", 1:3)),
  "UMAP1" = c(rep(x = 1, n = 3), rep(x = 2, n = 3), rep(x = 3, n = 3)),
  "UMAP2" = c(rep(x = 1:3, each = 3)),
  "c.type" = sample(c("CTA", "CTB", "CTC"), 9, replace = TRUE)
)
fitted_tissue <- tibble(
  "fitted_genes" = NA,
  "tissuemodel" = tibble(
    "bulk_id" = c("b1", "b1", "b1", "b2", "b2", "b2", "b3", "b3", "b3"),
    # "cell_id" = paste0("cell_", 1:9),
    "cell_id" = c(paste0("cell_", 1:3), paste0("cell_", 4:6), paste0("cell_", 1:3)),
    "weight" = c(1, 0.5, 0.25, 0.25, 0.5, 1, 0.3, 0.3, 0.3)
  )
)
bulkgroups <- c("b1" = "bgroup_1", "b2" = "bgroup_2", "b3" = "bgroup_1")
density <- density_distribution(
  scembedding = scembedding,
  x = "UMAP1", y = "UMAP2", scids = "colnames",
  fitted_tissue$tissuemodel,
  bulkgroups = bulkgroups
)

element_differential_density <- tissueResolver:::plot_element_differential_density(
  density,
  groupA = "bgroup_1",
  groupB = "bgroup_2",
  x = "x",
  y = "y",
  z = "z",
  group = "group"
)

print(ggplot() +
  element_differential_density +
  coord_fixed())
}
