% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{promote_matrix}
\alias{promote_matrix}
\title{Promotes a vector to a matrix. If the input is already a matrix, this matrix is returned.}
\usage{
promote_matrix(v)
}
\arguments{
\item{v}{matrix or vector}
}
\value{
promoted matrix
}
\description{
Promotes a vector to a matrix. If the input is already a matrix, this matrix is returned.
}
